Fixed Item Registry
Change registry method so each item has it's own 'item' rather then just diffrent nbt data.

Added a sort function to the CreativeMenuHandler. (if it doesn't sort the items, everything is jumbled up.)

Added in an ItemGrabber, can be used to get items for refrencing.

Added in an ItemJsonGenerator, for your new Item Needs!