fixed worldgen crash when loading up in a compiled environment.
while loading up a new world in an environment other than the development one the game would crash, calling an error on func_202792_ in certain layer files. this was due to it supering to a method that was after compiling considered abstract. replacing the super with the code the method contained resolved the issue presumably without changing functionality.