Fluid Tweaking
Still have an issue filling any machine in IE with IG fluids.

I suspect it may be linked to server or client side issues.

As, if you play in survival mode and attempt to fill a barrel, the image of the fluid bucket turns empty, but the barrel isn't filled.

If you then right click again, you'll find that the empty bucket will dispense the fluid it contained before the first attempt at filling the barrel.