Flasks, Fluids and Buckets!
Added in Flasks, for more... volatile liquids.
Added option for fluid material to use Flasks or Buckets, Not both.

NOTE: a better solution needs to be found for fluid storage/interactions, as it has a nasty case of allowing bucket usage for other mod storage stuff.

E.G: IE's barrels and tanks, return a bucket when a flask is emptied.