Foundation Work on Registration and Materials
MaterialPattern and it's children ItemPattern, FluidPattern and BlockPattern have been unified into MaterialFlags. While the functions related to Tags and ItemGroups will be moved to another class. This allows the use of a EnumSet as a way to flag the type of Item, Block or Fluid a Material Should Generate.

Registration has been changed to be in compliance with 1.18.2's Registration Changes